Improve compatibility of start.ps1 with GenisysPro
The original start.ps1 script only recognizes file "PocketMine-MP.phar".
It's ridiculous that we have a start.cmd which works with GenisysPro and
a start.ps1 which doesn't.
Therefore, I improved this PowerShell script to make it compatible with
GenisysPro artifact.
I still used "Genisys*.phar" and "Genisys.phar" as regexes to keep
consistency with start.cmd.