Changed default per-tick: to 1 (#32)
There is no benefit in having the per tick set to 4. Moreover, after testing I observed this consistently causes performance issues for players using tablets. Test was performed on a brand new setup. When per-tick: is set to 4, every 2 seconds while running/flying around (even in generated and/or cached chunks) there is a jerking action. When per-tick: is set to 1, game play is consistently smooth.