Redirect "supported version" section for convenience
They messed the "supported version" section up... only the section for
English is updated, leaving information for Chinese outdated. So I think
redirecting viewers to this section in English avoids this problem.
Obviously the version strings need not be translated, so the viewers
should understand, even if they are under the English section.